guixman script
changed 'full' to 'pull' and put the entire command for it in a seperate
script, should now only do icedove profile manager if the executable
location changed and kills existing instance and suppresses the output
of icedove so the reminder to do guixman os is still clear.

Think this is all I want from this script, we'll see how the workflow
feels but feels good to have a starting place.

Also the paths should probably be better managed but I didn't want all
the sub commands to be gexp lists so inserting varaibles would require
(string-append ...) etc which I didn't really want, this would be a
decent place to learn about custom syntax but that is very low on
priority list.