integrated slstatus into dwm and did dozonal
Also commited dwmsource.scm and pango_fix.scm, these are temporary until
I figure out proper scheme modules, primarily these are duplicate code
from other files so I definitely need to figure out how to import them

Main change is that I'm no longer using slstatus at all, basically all
the display is done differently or more efficiently. Main thing is that
it now can only redraw when the bar is visible and the clock is going to
update, but for the majority of the time it doesn't repeatedly read from
the battery info in the background every second, now it reads every 1/3
second when the bar is visible (based on update rate of clock) and
doesn't do any work when the bar is hidden.

the dwm_personal.diff is messy now, I wanted to put my files in a
seperate file which means I needed to add a diff flag to add new files
but that means the temporary files made by emacs and the config.h that
is a copy of the defaults is also copied over. I will need to figure out
how to deal with that, part of me is wondering about using tup.

Also you can ignore the changes to slstatus_personal.diff, I will be
removing it shortly along with other references to slstatus as I stole
all the code I needed from it and mine is better.