cleaned up home config
restructured package definitions so instead of being defined as
constants and then added to a master list later they are directly
declared in a big list of packages which reduces overhead.

Removed reference to specific libinput version, my patch has gone into
the mainstream and I was maintaining multiple versions of libinput
installed (found out from trying to use one transform for both dino and
libinput but it ended up recompiling dino from scratch because some of
it's dependencies use libinput) so it will be good to move to mainline
libinput.

Also added clearer markers for sections of home config and added -q to
the guix build command that could have caused dwm_personal.diff to go
very bad if the output included a build log instead of just the source
path.