finally added nix!
installed discord with nix and added nix stuff to guix config
specifically, .nix-defpath/default.nix is evaluated and used for nix-env
and when it resolves to a list of packages 'nix-env -i' just installs
those packages. I added 'nix-env -r -i' to some of the guixman commands
to remove existing packages and install the new ones, if there is no
change it doesn't output any details so it doesn't clutter the output.

Also idk what was going on with pango before but dwm and pkg-config
works perfectly well with the default pango, pango-1.90 exports
pango2.ac instead of pango.pc and xft.pc which is why guix shell was
giving me incorrect tests but I thought I tried what is now commited and
it didn't work but it is definitely working now.

guix updated to include nss-certs by default so that is now added to the
os, sorry for bundling commits together but I can't be bothered to
seperate them for a cleaner git history

Also I think I am getting close to getting the turris working but need
to figure out how to disable tests in graphene and I still don't even
know what depends on it. Also dozonal fonts but those are things to be
commited soon