partial keyfile
mostly added comments, discovered a bug that has lasted 24 months
which has it so that when you try to specify the keyfile to decrypt the
main partition in the initrd, it means the os building code doesn't
detect the encrypted partition since it doesn't match by equality the
expected mapping type and thus the code to mount it is just not put into
the grub.cfg.

I am mad about it, but at the same time it gives me something to
actually contribute and yell about where I know exactly why it is going
wrong but not sure what the suitable solution is.