updated dwm status function
added case for "Not charging" as that is what my battery reports instead
of "Full" also changed around the way it displays in certain cases to be
better.

I was having trouble getting the "Not charging" to display correctly, it
is currently only checking the first 10 characters of a 12 character
string but I changed that at the same time as using `x == 0` instead of
`!x` and using `%12[...]` instead of `%12s` so I'm not totally sure
which fixed it and I don't particularly care, what it does now is good
enough.