dunst service much better
Dunst service now properly handles symbol values and has more rigorous
handling of selectors, still requires the labels be symbols and the list
handling is all over the place. It still needs sanitization to validate
the fields when possible, also I need to figure out map-apply or
something because the unpacking of data from sub lists is getting really
annoying.