updated tupfile
I may have been incorrect about the local-file, it may have just not
been happy with the ../ relative location, so adding a symlink to the
aux-files folder to the root of the dotfiles is sufficient. In theory I
should just put the diff there but I don't want to deal with updating
guixman scripts I just want to figure out how to generate it from tup or
what ever build system I decide on.

the tup file is still not at all happy about the way the dependencies on
the go files is specified so it fails and the os command is failing due
to record types mismatch which I'm not sure how to fix.